Agente IA/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Werkzeug==2.2.3
python-dotenv==1.0.0
email-validator==2.0.0
typing-extensions==4.7.1
alembic==1.11.3
Mako==1.2.4
WTForms==3.0.1
dnspython==2.3.0
googlemaps==4.10.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
blinker==1.5
geopy==2.3.0
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.7
redis==5.0.1
python-redis-lock==4.0.0
APScheduler==3.10.4
chardet==5.2.0
PyPDF2==3.0.1
pdfplumber==0.11.7
Flask-Compress==1.13